refactor(infra): restructure Pulumi projects with src/ directory
Reorganize all soccer-stats infrastructure projects to use a consistent
src/ directory structure with smaller, logically organized modules.

Changes:
- soccer-stats-infra: Consolidate pulumi.ts into src/index.ts, add package.json
  for path alias matching
- soccer-stats-api-infra: Split monolithic pulumi.ts into modular files:
  - config.ts: Configuration variables
  - shared-infra.ts: Shared infrastructure imports
  - secrets.ts: Clerk secrets management
  - docker.ts: Docker image building
  - ecs.ts: Task definition and service
  - autoscaling.ts: Auto-scaling configuration
- soccer-stats-ui-infra: Split pulumi.ts into:
  - config.ts: Configuration variables
  - shared-infra.ts: Shared infrastructure imports
  - s3.ts: S3 bucket for static assets
  - cloudfront.ts: CloudFront distribution
  - sync.ts: File syncing to S3

Also fixes @nx/js:tsc path rewriting by ensuring target names match
across dependent projects and package.json names match path aliases.
