--- apps/soccer-stats/api-infra/index.ts
@@ -1,2 +1,0 @@
-import './pulumi';
-export * from './pulumi';
