+++ apps/soccer-stats/ui-infra/Pulumi.yaml
@@ -1,6 +1,6 @@
 name: soccer-stats-ui-infra
 description: S3 and CloudFront infrastructure for soccer-stats UI
-main: ../../../dist/apps/soccer-stats/ui-infra/pulumi.js
+main: ../../../dist/apps/soccer-stats/ui-infra/src/index.js
 runtime:
   name: nodejs
   options:
