+++ apps/soccer-stats/ui-infra/src/stack-reference.ts
