+++ libs/soccer-stats/infra/package.json
@@ -1,7 +1,4 @@
 {
   "name": "@garage/soccer-stats/infra",
-  "version": "0.0.1",
-  "type": "module",
-  "main": "./src/index.ts",
-  "types": "./src/index.ts"
+  "version": "0.0.1"
 }
